refactor: extract shared browse primitives into browse/shared
