+++ src/features/browse/shared/StructureView.css
