+++ src/features/browse/shared/binaryCell.ts
