+++ src/features/browse/shared/highlightSql.ts
